+++ packages/terra-base/tests/nightwatch/NoTranslationsLoadedStringBase.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import Base from '../../lib/Base';
+
+const BaseNoTranslationsLoaded = () => {
+  const noTranslations = (
+    <Base locale="gabcdef" translationsLoadingPlaceholder={'No Translations String'}>
+      <div>Translations</div>
+    </Base>
+  );
+  return noTranslations;
+};
+
+export default BaseNoTranslationsLoaded;