+++ packages/terra-signature/tests/jest/Signature.test.jsx
@@ -6,23 +6,23 @@ describe('Signature', () => {
 
   // Snapshot Tests
   it('should render a default component', () => {
-    const signature = shallow(defaultRender);
+    const signature = render(defaultRender);
     expect(signature).toMatchSnapshot();
   });
 
   it('should receive line width prop', () => {
-    const signature = shallow(<Signature lineWidth={Signature.Opts.Width.EXTRA_FINE} />);
+    const signature = render(<Signature lineWidth={Signature.Opts.Width.EXTRA_FINE} />);
     expect(signature).toMatchSnapshot();
   });
 
   it('should receive line segments prop', () => {
-    const signature = shallow(<Signature lineSegments={[]} />);
+    const signature = render(<Signature lineSegments={[]} />);
     expect(signature).toMatchSnapshot();
   });
 
   it('should have custom props', () => {
     const singleLine = <Signature id="foo" />;
-    const signature = shallow(singleLine);
+    const signature = render(singleLine);
     expect(signature).toMatchSnapshot();
   });
 });