+++ packages/terra-site/src/TestLinks.jsx
@@ -22,6 +22,8 @@ const TestLinks = () => (
       <li><Link to="/tests/dynamic-grid-tests">Dynamic Grid Tests</Link></li>
       <li><Link to="/tests/embedded-content-consumer-tests">EmbeddedContentConsumer Tests</Link></li>
       <li><Link to="/tests/form-tests">Form Tests</Link></li>
+      <li><Link to="/tests/form-field-tests">Form Field Tests</Link></li>
+      <li><Link to="/tests/form-textarea-tests">Form Textarea Tests</Link></li>
       <li><Link to="/tests/grid-tests">Grid Tests</Link></li>
       <li><Link to="/tests/heading-tests">Heading Tests</Link></li>
       <li><Link to="/tests/hookshot-tests">Hookshot Tests</Link></li>
@@ -38,6 +40,7 @@ const TestLinks = () => (
       <li><Link to="/tests/modal-tests">Modal Tests</Link></li>
       <li><Link to="/tests/modal-manager-tests">Modal Manager Tests</Link></li>
       <li><Link to="/tests/multi-select-list-tests">MultiSelectList Tests</Link></li>
+      <li><Link to="/tests/multi-select-table-tests">MultiSelectTable Tests</Link></li>
       <li><Link to="/tests/overlay-tests">Overlay Tests</Link></li>
       <li><Link to="/tests/overlay-container-tests">OverlayContainer Tests</Link></li>
       <li><Link to="/tests/popup-tests">Popup Tests</Link></li>
@@ -45,11 +48,13 @@ const TestLinks = () => (
       <li><Link to="/tests/progress-bar-tests">ProgressBar Tests</Link></li>
       <li><Link to="/tests/props-table-tests">Props Table Tests</Link></li>
       <li><Link to="/tests/responsive-element-tests">ResponsiveElement Tests</Link></li>
-      <li><Link to="/tests/search-field-tests">Search Field</Link></li>
+      <li><Link to="/tests/search-field-tests">Search Field Tests</Link></li>
+      <li><Link to="/tests/select-tests">Select Tests</Link></li>
+      <li><Link to="/tests/select-option-tests">Select.Option Tests</Link></li>
       <li><Link to="/tests/selectable-list-tests">Selectable List Tests</Link></li>
       <li><Link to="/tests/signature-tests">Signature Tests</Link></li>
-      <li><Link to="/tests/selectable-table-tests">SelectableTable Tests</Link></li>
       <li><Link to="/tests/single-select-list-tests">SingleSelectList Tests</Link></li>
+      <li><Link to="/tests/selectable-table-tests">SelectableTable Tests</Link></li>
       <li><Link to="/tests/single-select-table-tests">SingleSelectTable Tests</Link></li>
       <li><Link to="/tests/slide-group-tests">SlideGroup Tests</Link></li>
       <li><Link to="/tests/slide-panel-tests">SlidePanel Tests</Link></li>