+++ packages/terra-site/src/examples/search-field/SearchFieldDisabled.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import SearchFieldExampleTemplate from './SearchFieldExampleTemplate';
+
+const SearchFieldDisabled = () => (
+  <SearchFieldExampleTemplate
+    isDisabled
+  />
+);
+
+export default SearchFieldDisabled;