+++ packages/terra-base/CHANGELOG.md
@@ -3,6 +3,13 @@ Changelog
 
 Unreleased
 ----------
+### Changed
+* componentWillReceiveProps will now compare locale instead of props.
+
+2.7.0 - (November 16, 2017)
+------------------
+### Changed
+* Minor version bump
 
 2.6.0 - (October 6, 2017)
 ------------------