+++ packages/terra-button/CHANGELOG.md
@@ -3,6 +3,13 @@ Changelog
 
 Unreleased
 ----------
+### Changed
+* Corrected typo in `terra-button` css comment. 
+
+1.13.0 - (November 16, 2017)
+------------------
+### Changed
+* Removed unnecessary `font-family` rule.
 
 1.12.0 - (October 31, 2017)
 ------------------