+++ packages/terra-date-picker/CHANGELOG.md
@@ -3,6 +3,18 @@ Changelog
 
 Unreleased
 ----------
+### Changed
+* Removed onInputFocus, onCalendarButtonClick and shouldShowPicker props and pass them down as custom props.
+
+1.16.0 - (November 16, 2017)
+------------------
+### Added
+* Added onInputFocus and onCalendarButtonClick props to handle unacceptable default dates.
+
+1.15.0 - (November 7, 2017)
+------------------
+### Changed
+* Minor version bump
 
 1.14.0 - (October 31, 2017)
 ------------------