+++ packages/terra-date-time-picker/CHANGELOG.md
@@ -4,6 +4,16 @@ ChangeLog
 Unreleased
 ----------
 
+1.3.0 - (November 16, 2017)
+------------------
+### Added
+* Handle unacceptable default dates.
+
+1.2.0 - (November 7, 2017)
+------------------
+### Changed
+* Minor version bump
+
 1.1.0 - (October 31, 2017)
 ------------------
 ### Changed