+++ packages/terra-list/CHANGELOG.md
@@ -4,6 +4,16 @@ Changelog
 Unreleased
 ----------
 
+1.16.0 - (November 16, 2017)
+------------------
+### Fixed
+* Chevron positioning
+
+1.15.0 - (November 7, 2017)
+------------------
+### Changed
+* Minor version bump
+
 1.14.0 - (October 31, 2017)
 ------------------
 ### Changed