+++ packages/terra-menu/CHANGELOG.md
@@ -4,6 +4,16 @@ Changelog
 Unreleased
 ----------
 
+1.8.0 - (November 16, 2017)
+------------------
+### Changed
+* Minor version bump
+
+1.7.0 - (November 7, 2017)
+------------------
+### Changed
+* Minor version bump
+
 1.6.0 - (October 31, 2017)
 ------------------
 ### Added