+++ packages/terra-signature/CHANGELOG.md
@@ -4,6 +4,17 @@ Changelog
 Unreleased
 ----------
 
+0.6.0 - (November 16, 2017)
+------------------
+### Changed
+* Minor version bump
+
+0.5.0 - (November 7, 2017)
+------------------
+### Changed
+* Removing event listeners from the right element during component unmount
+* Fix onChange, it was returning the last segment instead of the array of the whole last line segments
+
 0.4.0 - (October 24, 2017)
 ------------------
 ### Changed