+++ packages/terra-site/CHANGELOG.md
@@ -4,6 +4,18 @@ Changelog
 Unreleased
 ----------
 
+1.18.0 - (November 16, 2017)
+------------------
+### Added
+* Add terra-form-textarea components
+* Add terra-form-field components
+* Add terra-mock-theme styles
+
+1.17.0 - (November 7, 2017)
+------------------
+### Changed
+* Minor version bump
+
 1.16.0 - (October 31, 2017)
 ------------------
 ### Changed