+++ packages/terra-time-input/CHANGELOG.md
@@ -3,6 +3,13 @@ Changelog
 
 Unreleased
 ----------
+### Changed
+* Removed onInputFocus prop and pass it down as a custom prop.
+
+1.14.0 - (November 16, 2017)
+------------------
+### Added
+* Added onInputFocus prop.
 
 1.13.0 - (October 24, 2017)
 ------------------